save wasm dynamic library temp files if in DEBUG mode, as with others

+++ emcc.py
@@ -2074,8 +2074,9 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
             wso = shared.WebAssembly.make_shared_library(js_target, wasm_binary_target)
             # replace the wasm binary output with the dynamic library. TODO: use a specific suffix for such files?
             shutil.move(wso, wasm_binary_target)
-            os.unlink(js_target) # we don't need the js, it can just confuse
-            os.unlink(asm_target) # we don't need the asm.js, it can just confuse
+            if not DEBUG:
+              os.unlink(js_target) # we don't need the js, it can just confuse
+              os.unlink(asm_target) # we don't need the asm.js, it can just confuse
             sys.exit(0) # and we are done.
 
       # If we were asked to also generate HTML, do that
